requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1
PyMySQL==1.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
Jinja2==3.1.2
SQLAlchemy>=1.4.0
cryptography==41.0.7
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
python-docx
weasyprint 